fix(dataclass): rez execution on UNIX system
problem reported by Luke on slack where rez-env was ignoring the list of package provided to it.
This was due to the combination of shell=True and command being a sequence, all ran on an UNIX system.

To address this issue I removed the shell argument and retrieve the path to the rez-env executable manually before.